refactor: Restructure AccommodationDetail component with modular subcomponents for improved readability and maintainability
